Meteorolog/requirements.txt
apscheduler
numpy
pydantic
pydantic-settings
pyyaml
rich
requests
sqlalchemy
pyserial
smbus2
psutil